+++ test/GIL.jl
@@ -1,8 +1,8 @@
-@testitem "release and lock" begin
-    # This calls Python's time.sleep(1) twice concurrently. Since sleep() releases the
+@testitem "unlock and lock" begin
+    # This calls Python's time.sleep(1) twice concurrently. Since sleep() unlocks the
     # GIL, these can happen in parallel if Julia has at least 2 threads.
     function threaded_sleep()
-        PythonCall.GIL.release() do
+        PythonCall.GIL.unlock() do
             Threads.@threads for i = 1:2
                 PythonCall.GIL.lock() do
                     pyimport("time").sleep(1)
@@ -20,11 +20,11 @@
     end
 end
 
-@testitem "@release and @lock" begin
-    # This calls Python's time.sleep(1) twice concurrently. Since sleep() releases the
+@testitem "@unlock and @lock" begin
+    # This calls Python's time.sleep(1) twice concurrently. Since sleep() unlocks the
     # GIL, these can happen in parallel if Julia has at least 2 threads.
     function threaded_sleep()
-        PythonCall.GIL.@release Threads.@threads for i = 1:2
+        PythonCall.GIL.@unlock Threads.@threads for i = 1:2
             PythonCall.GIL.@lock pyimport("time").sleep(1)
         end
     end